+++ osidb/tests/test_flaw.py
@@ -187,7 +187,6 @@ class TestFlaw:
             "name",
             "company",
             from_upstream=True,
-            meta_attr={"foo": "bar"},
             acl_read=self.acl_read,
             acl_write=self.acl_write,
         )